added cacheadapter interface and memorycache of notification.
this patch included cacheadapter and using the memorycache.
previosuly, we used the only observeid(unique id).
currently, we use the ip in OCDevaddr Struct.

Reviewed-on: https://gerrit.iotivity.org/gerrit/8191
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ service/notification/src/provider/NSProviderDiscovery.c
@@ -65,7 +65,7 @@ void * NSDiscoverySchedule(void * ptr)
             NSTask *node = NSHeadMsg[DISCOVERY_SCHEDULER];
             NSHeadMsg[DISCOVERY_SCHEDULER] = node->nextTask;
 
-            switch (node->taskType)
+            switch ((int)node->taskType)
             {
                 case TASK_START_PRESENCE:
                     NSStartPresence();
@@ -77,6 +77,7 @@ void * NSDiscoverySchedule(void * ptr)
                     NSRegisterResource();
                     break;
             }
+
             OICFree(node);
         }
